requirements.txt
openai-agents>=0.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic>=2.0.0
google-search-results>=2.4.2
trafilatura>=1.6.0
readability-lxml>=0.8.1
pymupdf>=1.23.0
playwright>=1.40.0
faiss-cpu>=1.7.4
numpy>=1.24.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-genai>=0.3.0
